refactor: formatTimeファンクションをclassにする
